Open open/close fetch_history function in the core loader
Currently, it is the task's job to open/close fetch_history.
After discussion, we must keep the task as minimal as possible.
So it comes down to let the loader do that in the `load` function call
as well.

This is also a step towards merging swh.loader.git.base.BaseLoader.