Ignore local history when we miss an object
Summary:
The git protocol will not allow us to "fill the gaps" if we just filter out the
missing HEADs: if we have data that's been created more recently, the packfile
returned by the server will be empty.

This is not the current T1211 but this addresses one of the earliest failure
cases of the git loader, where one of the branches wouldn't load and where
further visits would yield a truncated history because of the current
incremental behavior.

Test Plan:
Loading an origin manually, then removing one of the HEADs, then
loading again fills the gap.

Reviewers: #reviewers!

Differential Revision: https://forge.softwareheritage.org/D456