package.loader: Reference a snapshot on partial visit
Prior to this commit, partial visit would not reference a partial snapshot.
Thus further visits on the same origin would not benefit from the snapshot's
already seen artifacts. This now improve on such behavior to reference such
partial snapshot.

For the deposit's case though, we want to fail immediately when a deposit id is
unknown. This avoids creating noises in the storage.