nixguix: add the integrity attribute in release metadata
The integrity attribute is also used by the incremental loading
feature.

Before, the url was used for incremental loading: if an url is already
existing in the last snapshot, this url is not downloaded again.

Using the url as key for this cache has several drawbacks:
- if the upstream source is upgraded in place, the snapshot will reuse
  the previous version of the upstream source.
- when the loader supports url mirrors, using the url as branch name will
  be confusing (we won't know which url to use)

The integrity attribute is now used as key (instead of the url).

Moreover, this integrity attribute will be useful to get artifacts
from SWH once a metadata API will be available.