git.loader: fix failing origin visit update step
due to uninitialized state variables

Currently, a failure happens for some reason. As some internal state variables
are not initialized (they are in fetch_data but not in __init__), this
makes the get_snapshot method fails because of it. Then that's making the
visit update step fails.

Adding the proper initialization for those variables in the __init__ method
should fix it. That will allow the failed state of the visit to be updated
properly in the storage. Associating an empty snapshot to the failed visit.

This should fix issues like [1]

https://sentry.softwareheritage.org/share/issue/ace9143c0f6244b49b18cca49a5aa809/