qdeposit.loader: Fix committer date appropriately
This adds another test scenario where committer and date are different. This is
to further check the dates are set correctly in the revision. In the previous
test data, the committer date and author were the same so no hash change were
detected.

This also adds a note about the date provenance. The ones used to fill in the
revision date fields.

Related to T2368