nixguix: Override known_artifacts to filter out evaluation branch
As that branch is a special case of targetting nixpkgs, this does not have the
same and expected metadata the other revisions do. Thus crashing the loader
when this one tries to resolve from a snapshot the already seen artifacts (by
looking into the metadata).

This commits fixes that behavior by excluding the "evaluation" branch from the
list of revisions to resolve.

Related to T2411