Only use the previous (base) snapshot as reference for incremental loads
The old behavior would find any object with a matching sha1_git from the archive
to use as a reference point in the generated snapshot. While there's no known
sha1_git collisions in the wild, this behavior is inconsistent with other
loaders.

This removes the only reason to keep the `object_find_by_sha1_git` api in storage.