Adapt the git loader to swh-model >0.4
in which extra_headers is now a top-level attribute of a Revision.

Add tests that actually include having extra_header in loaded git revisions.
For this, the fast-import based dump of the test git repo is replaced by
a bundle dump (required to ensure things like gpg signatures are kept).
A few git revisions including gpg signature, encoding and mergetag
headers are also added to the git repo.