core.loader: Log information about origin currently being ingested
Prior to this commit, we "info" logged no meaningful information about the
origin being loaded. With the following commit, we will be able to tell what
origin/visit-type is happening.

This also adds some extra assertion failure message to explicit why the
self.origin and self.visit should be set and where to look if not.