Revert "setup: Fix long description content type to rst"
This reverts commit dbb18628f16d2cf68d8d71eb1c3cc6426017f3c4.

This creates issues when uploading to pypi [1].

The type x-rst is not posing problem for example in swh.core. But in that repository the
readme is not a symlink... I'm not investigating this right now so a simple revert
should do. My focus is on deploying the opam loader on staging for now.

[1] https://jenkins.softwareheritage.org/view/swh-draft/job/DLDBASE/job/pypi-upload/111/console