Improve store_data implem to allow multiple calls with partial visit
This introduces a `create_partial_visit` flag to the `store_data` method call which
allows to write partial visit with the current snapshot built during one iteration of
the ingestion loop.

The final loop of the visit will not create a partial visit. It will create the last
visit with status "full" targeting the final snapshot as before.

The main difference is that an ingestion is actually more verbose in terms of
origin_visit_status. Which in turn allows to be more incremental in subsequent visit of
the same origin.

Related to T3625