git: Ingest ordered tags then ordered branches references
This should allow to gradually ingest the git history. If we don't deal with
{master/main/HEAD} references first, it sounds highly plausible we won't have a mostly
full connected graph at first (except for repositories with no tags). Started from tag
v0.1.0, v0.2.0, etc...

If we couple this with the option to ingest repository with multiple packfiles, this
should allow to ingest more incrementally large repositories (as it creates partial
snapshots at each fetch/store loop).

Related to T3625