Consider unauthorized access to origin as a not found visit status
This currently fails the origin visit and update the visit status to 'failed'.

This got listed by listers but current access to such origin is actually private, it'd
probably make sense to make the status of the visit as not_found instead.

This takes care of the most frequent issue so (460k) [1].

[1] https://sentry.softwareheritage.org/share/issue/3a3663f8cc424a48999af28728152ef0/