Use all base snapshots in determine_wants()
Before this commit, determine_wants() used the origin's last snapshot
if any, or the closest parent's snapshot if not.

However, we noticed that many repositories that are very slow to load
are forks that were already visited, but their owner rebased it on the
parent since the last visit, causing potentially many commits to be
added to the origin.

This ensures we do not needlessly fetch these new commits when we
already loaded the parent.