Use all snapshots to compute base_snapshot_reverse_branches
base_snapshot_reverse_branches needs to contain all objects that
may be a snapshot target that the remote did not send to us.

Because we now use all snapshots to build the "have" list, such
targets include all targets of snapshots of "parent" origins,
not just the previous snapshot of the current origin.

This typically happens when a forge-fork pull branches from its
parent.

Resolves Sentry issue [[ https://sentry.softwareheritage.org/share/issue/1c08f5d764e7494e83ba254dc47f17af/ | SWH-LOADER-GIT-102 ]]