Only traverse the tag/commit graph
On linux.git, the full traversal on trees and blobs uses over 50GB of RAM;
while it is less than 1GB on tag/commits only.

After running the traversal:

1. build a pack index
2. iterate over commit/tags in the packfile
3. for each commit/tags we have not seen while traversing (ie. that are
   not reachable from known snapshots), recursively load the tree and
   blobs they reference

In terms of benchmarks (on my machine, and using dulwich 0.20.43 +
in-memory overlay storage on top of production):

* 2h with this commit
* 2.5h with a variant of this commit, using the on-disk indexes
  (instead of MemoryPackIndex)
* 2.75h with the previous algorithm

And this is despite this commit being naive, in that will repeatedly try
to load the same trees and blobs when they were already.
A future commit may improve this; eg. by marking what trees and blobs we
loaded in the pack index.