loader: Add statsd metric to count number of resolved external refs
Add a statsd counter to report the total number of git objects per type
resolved from the archive but also the total number of objects that could
not be resolved.

Ensure counters are incremented exactly once per git object as the
content of the pack file is iterated four times, one per git object
type.