loader: Ensure NotFound exception is reraised when caught
Previous commit modified the dumb.check_protocol function to raise an
HTTPError exception when the request to check dumb protocol support
failed. As NotFound exception inherits from ValueError, the code for
checking dumb protocol support was executed even when a repository was
not found. So an HTTPError exception was raised with a 404 status code
and the NotFound exception was no longer propagated to the base loader
class, resulting in a failed visit status instead of a not_found one.