directory: Use treeless cloning to make the cloning step faster
Inspired from the pip cloning step [1]. This makes the cloning steps only fetch the
commit information and the tree at the current heads. Then a subsequent
switch (checkout) retrieves the tree at the reference we want. In effect, this retrieves
way faster the necessary tree needed to ingest the repository.

[1] it uses a blobless cloning though.