dumb: Fix streaming of HTTP responses
When using the requests library to perform HTTP requests, if responses
need to be streamed the stream parameter must be set to True to ensure
content is downloaded by chunks.

Previously, a whole HTTP response was cached in memory which could lead
to OOM errors when dealing with a repository with large pack files.