test_loader: Fix implementation of test_loader_with_ref_delta_in_pack
Previous implementation was building an invalid pack file with REF_DELTA
object types as it was using the new object to deltify as the base of the
delta.

This was leading to errors and undefined behavior after building an index
for such a pack file as the deltified objects could not be properly resolved
by dulwich (observed by stsp while working on git loader improvements).

The bases for deltified objects are now objects that were previously loaded
into the archive.

Tag objects produced in that test are also ensured to be valid.